--- a/src/crimson/os/seastore/segment_cleaner.h
+++ b/src/crimson/os/seastore/segment_cleaner.h
@@ -516,10 +516,10 @@ public:
 	  12,   // target_journal_segments
 	  16,   // max_journal_segments
 	  2,	// target_backref_inflight_segments
-	  .9,   // available_ratio_gc_max
-	  .2,   // available_ratio_hard_limit
-	  .8,   // reclaim_ratio_hard_limit
-	  .6,   // reclaim_ratio_gc_threshold
+	  .1,   // available_ratio_gc_max
+	  .05,  // available_ratio_hard_limit
+	  .9,   // reclaim_ratio_hard_limit
+	  .1,   // reclaim_ratio_gc_threshold
 	  1<<20,// reclaim_bytes_per_cycle
 	  1<<17,// rewrite_dirty_bytes_per_cycle
 	  1<<24 // rewrite_backref_bytes_per_cycle