It should be generically better to delay reclaim as late as possible, so
that:
* unalive/unavailable ratio can become higher to reduce reclaim efforts, see before[^1] to after[^2], and more effective in case of before[^5] to after[^6] from https://github.com/ceph/ceph/pull/46436, down to ~1/3 reclaim efforts;
* less conflicts between mutate and reclaim transactions, see before[^3] to after[^4], down to ~1% conflicts from reclaim transactions;

Please note,
* in the non-generational version[^1][^2][^3][^4], reclaiming is too slow so that unavailable/total ratio will reach the original hard limit 80%, **not** significantly lower than the tuned 95% limit, so that the end performance is **not** significantly impacted, from 11.9MB/s to 11.95MB/s.
* in the generational version[^5][^6] from https://github.com/ceph/ceph/pull/46436, due to the improved bimodal distribution of segment utilizations, reclaiming is much faster so that the unavailable/total ratio ends up to ~45%, significantly lower than the tuned 95% limit, so that the end performance is improved from 9.95MB/s to 12.31MB/s.

Another change is `crimson/os/seastore/transaction_manager: set to test mode under debug build`, please refer to its commit message.

[^1]: ![before-waf](https://user-images.githubusercontent.com/7736006/171078335-bb381415-f72c-493b-b2e4-1d1aebb05bfe.png)
[^2]: ![after-waf](https://user-images.githubusercontent.com/7736006/171078388-1a8298ab-745a-4dcf-a539-4fd074ccf628.png)
[^3]: ![before-conflicts](https://user-images.githubusercontent.com/7736006/171078508-0233aec9-e2ce-4eb8-b6af-67213504dff3.png)
[^4]: ![after-conflicts](https://user-images.githubusercontent.com/7736006/171078541-d9c5b080-870f-4ea8-8d16-ea147d868f57.png)
[^5]: ![before2-waf](https://user-images.githubusercontent.com/7736006/171078581-9884a20a-50f4-451f-93d2-f995945d6686.png)
[^6]: ![after2-waf](https://user-images.githubusercontent.com/7736006/171078629-a42e85d5-15f4-4b95-81ea-43ca37be8e70.png)




<!--
  - Please give your pull request a title like

      [component]: [short description]

  - Please use this format for each git commit message:

      [component]: [short description]

      [A longer multiline description]

      Fixes: [ticket URL on tracker.ceph.com, create one if necessary]
      Signed-off-by: [Your Name] <[your email]>

    For examples, use "git log".
-->

## Contribution Guidelines
- To sign and title your commits, please refer to [Submitting Patches to Ceph](https://github.com/ceph/ceph/blob/master/SubmittingPatches.rst).

- If you are submitting a fix for a stable branch (e.g. "pacific"), please refer to [Submitting Patches to Ceph - Backports](https://github.com/ceph/ceph/blob/master/SubmittingPatches-backports.rst) for the proper workflow.

## Checklist
- Tracker (select at least one)
  - [ ] References tracker ticket
  - [ ] Very recent bug; references commit where it was introduced
  - [ ] New feature (ticket optional)
  - [ ] Doc update (no ticket needed)
  - [ ] Code cleanup (no ticket needed)
- Component impact
  - [ ] Affects [Dashboard](https://tracker.ceph.com/projects/dashboard/issues/new), opened tracker ticket
  - [ ] Affects [Orchestrator](https://tracker.ceph.com/projects/orchestrator/issues/new), opened tracker ticket
  - [ ] No impact that needs to be tracked
- Documentation (select at least one)
  - [ ] Updates relevant documentation
  - [ ] No doc update is appropriate
- Tests (select at least one)
  - [ ] Includes [unit test(s)](https://docs.ceph.com/en/latest/dev/developer_guide/tests-unit-tests/)
  - [ ] Includes [integration test(s)](https://docs.ceph.com/en/latest/dev/developer_guide/testing_integration_tests/)
  - [ ] Includes bug reproducer
  - [ ] No tests

<details>
<summary>Show available Jenkins commands</summary>

- `jenkins retest this please`
- `jenkins test classic perf`
- `jenkins test crimson perf`
- `jenkins test signed`
- `jenkins test make check`
- `jenkins test make check arm64`
- `jenkins test submodules`
- `jenkins test dashboard`
- `jenkins test dashboard cephadm`
- `jenkins test api`
- `jenkins test docs`
- `jenkins render docs`
- `jenkins test ceph-volume all`
- `jenkins test ceph-volume tox`
- `jenkins test windows`
</details>